Add stop_at_round option to end the fight after X rounds
Example: stop_at_round=1 will end the fight after 1 round of combat, to
simulate the damage (ignore the win percentage in that case). Which can
be helpful to simulate how bad a fight can be assuming a retreat ASAP.

It does not take into account the extra round of damage if the retreater
goes second.